Remove accessibility permission and related code
This commit removes the `AccessibilityPermission` class and its usage. It also updates the AndroidManifest.xml to remove the `TrackerAccessibilityService` and related permissions. Additionally, it removes the media projection-related code from the Android dependency injection module.

The build.gradle.kts file is updated to add the MapKit framework for iOS and the TomTom map key for Android.

The AuthNavGraph.kt file has a minor formatting change.